refactor: move linkify pipe from features/pipes to shared/pipes
